Cover eth.db.* in API docs
