+++ docs/api/db/api.db.atomic.rst
@@ -1,0 +1,13 @@
+Atomic
+======
+
+AtomicDB
+~~~~~~~~
+
+.. autoclass:: eth.db.atomic.AtomicDB
+  :members:
+
+
+.. autoclass:: eth.db.atomic.AtomicDBWriteBatch
+  :members:
+
