+++ docs/api/db/api.db.batch.rst
@@ -1,0 +1,8 @@
+Batch
+=====
+
+BatchDB
+~~~~~~~
+
+.. autoclass:: eth.db.batch.BatchDB
+  :members:
