+++ docs/api/db/api.db.cache.rst
@@ -1,0 +1,8 @@
+Cache
+=====
+
+CacheDB
+~~~~~~~
+
+.. autoclass:: eth.db.cache.CacheDB
+  :members:
