+++ docs/api/db/api.db.diff.rst
@@ -1,0 +1,22 @@
+DBDiff
+======
+
+DBDiff
+~~~~~~
+
+.. autoclass:: eth.db.diff.DBDiff
+  :members:
+
+DBDiffTracker
+~~~~~~~~~~~~~
+
+.. autoclass:: eth.db.diff.DBDiffTracker
+  :members:
+
+DiffMissingError
+~~~~~~~~~~~~~~~~
+
+.. autoclass:: eth.db.diff.DiffMissingError
+  :members:
+
+
