+++ docs/api/db/api.db.header.rst
@@ -1,0 +1,8 @@
+Header
+======
+
+HeaderDB
+~~~~~~~~
+
+.. autoclass:: eth.db.header.HeaderDB
+  :members:
