+++ docs/api/db/api.db.schema.rst
@@ -1,0 +1,8 @@
+Schema
+======
+
+SchemaV1
+--------
+
+.. autoclass:: eth.db.schema.SchemaV1
+  :members:
