+++ docs/api/db/api.db.storage.rst
@@ -1,0 +1,14 @@
+Storage
+=======
+
+AccountStorageDB
+----------------
+
+.. autoclass:: eth.db.storage.AccountStorageDB
+  :members:
+
+StorageLookup
+-------------
+
+.. autoclass:: eth.db.storage.StorageLookup
+  :members:
