+++ eth/db/schema.py
@@ -1,34 +1,12 @@
-from abc import ABC, abstractmethod
-
 from eth_typing import (
     BlockNumber,
     Hash32,
 )
 
-
-class BaseSchema(ABC):
-    @staticmethod
-    @abstractmethod
-    def make_canonical_head_hash_lookup_key() -> bytes:
-        raise NotImplementedError('Must be implemented by subclasses')
-
-    @staticmethod
-    @abstractmethod
-    def make_block_number_to_hash_lookup_key(block_number: BlockNumber) -> bytes:
-        raise NotImplementedError('Must be implemented by subclasses')
-
-    @staticmethod
-    @abstractmethod
-    def make_block_hash_to_score_lookup_key(block_hash: Hash32) -> bytes:
-        raise NotImplementedError('Must be implemented by subclasses')
-
-    @staticmethod
-    @abstractmethod
-    def make_transaction_hash_to_block_lookup_key(transaction_hash: Hash32) -> bytes:
-        raise NotImplementedError('Must be implemented by subclasses')
+from eth.abc import SchemaAPI
 
 
-class SchemaV1(BaseSchema):
+class SchemaV1(SchemaAPI):
     @staticmethod
     def make_canonical_head_hash_lookup_key() -> bytes:
         return b'v1:canonical_head_hash'
